refactor: simplify category caching logic
Removes the `storedLanguage` parameter from `getMovieCategories` and `getTvShowCategories` in `CategoryLocalSource` and its implementation.

This simplifies the caching logic in `MovieRepositoryImpl` and `TvShowRepositoryImpl` by renaming `cacheMovieCategories` and `cacheTvShowCategories` to `cacheMovieCategoriesIfNotCached` and `cacheTvShowCategoriesIfNotCached` respectively, and removing redundant language preference checks.

Additionally, a typo in `TvShowRepositoryImpl` was corrected where `getMovieCategories` was called instead of `getTvShowCategories` and `toLocalDtoList` was called instead of `toLocalTvShowCategoryDtoList`.